Continued splitting commands in separate files
